+++ Common/src/main/java/com/nyfaria/spookybats/entity/HerobrineBat.java
@@ -21,7 +21,7 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.ServerLevelAccessor;
 import org.jetbrains.annotations.Nullable;
 
-public class HerobrineBat extends MonsterBat {
+public class HerobrineBat extends  MonsterBat {
 	public HerobrineBat(EntityType<? extends SpookyBat> entityType, Level level) {
 		super(entityType, level);
 		this.xpReward = 7;