+++ Common/src/main/java/com/nyfaria/spookybats/entity/api/Master.java
@@ -1,0 +1,8 @@
+package com.nyfaria.spookybats.entity.api;
+
+public interface Master {
+    void minionDeath();
+    void setMinionCount(int count);
+    int getMinionCount();
+
+}