+++ Fabric/src/main/java/com/nyfaria/spookybats/SpookyBats.java
@@ -4,10 +4,7 @@ import com.nyfaria.spookybats.init.EntityInit;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.biome.v1.BiomeModifications;
 import net.fabricmc.fabric.api.biome.v1.BiomeSelectors;
-import net.fabricmc.fabric.api.object.builder.v1.block.type.BlockSetTypeRegistry;
 import net.fabricmc.fabric.api.object.builder.v1.entity.FabricDefaultAttributeRegistry;
-import net.minecraft.client.renderer.ItemBlockRenderTypes;
-import net.minecraft.tags.BiomeTags;
 import net.minecraft.world.entity.MobCategory;
 import net.minecraft.world.level.biome.Biomes;
 
@@ -19,18 +16,24 @@ public class SpookyBats implements ModInitializer {
         EntityInit.attributeSuppliers.forEach(
                 p -> FabricDefaultAttributeRegistry.register(p.entityTypeSupplier().get(), p.factory().get().build())
         );
-        BiomeModifications.addSpawn(BiomeSelectors.foundInOverworld(), MobCategory.MONSTER, EntityInit.PUMPKIN_BAT.get(), 5, 1, 3);
-        BiomeModifications.addSpawn(BiomeSelectors.foundInOverworld(), MobCategory.MONSTER, EntityInit.CREEPER_BAT.get(), 5, 1, 3);
-        BiomeModifications.addSpawn(BiomeSelectors.includeByKey(Biomes.SWAMP,Biomes.MANGROVE_SWAMP), MobCategory.MONSTER, EntityInit.WITCH_BAT.get(), 5, 1, 3);
-        BiomeModifications.addSpawn(BiomeSelectors.foundInOverworld(), MobCategory.MONSTER, EntityInit.STEVE_BAT.get(), 5, 1, 3);
-        BiomeModifications.addSpawn(BiomeSelectors.foundInOverworld(), MobCategory.MONSTER, EntityInit.ALEX_BAT.get(), 5, 1, 3);
-        BiomeModifications.addSpawn(BiomeSelectors.foundInOverworld(), MobCategory.MONSTER, EntityInit.HEROBRINE_BAT.get(), 5, 1, 3);
-        BiomeModifications.addSpawn(BiomeSelectors.foundInOverworld(), MobCategory.MONSTER, EntityInit.SKELETON_BAT.get(), 5, 1, 3);
-        BiomeModifications.addSpawn(BiomeSelectors.foundInTheNether(), MobCategory.MONSTER, EntityInit.WITHER_SKELETON_BAT.get(), 5, 1, 3);
-        BiomeModifications.addSpawn(BiomeSelectors.foundInTheNether(), MobCategory.MONSTER, EntityInit.UNDEAD_BAT.get(), 5, 1, 3);
-        BiomeModifications.addSpawn(BiomeSelectors.foundInOverworld(), MobCategory.MONSTER, EntityInit.PLAYER_BAT.get(), 5, 1, 3);
-        BiomeModifications.addSpawn(BiomeSelectors.includeByKey(Biomes.DEEP_DARK), MobCategory.MONSTER, EntityInit.SCULK_BAT.get(), 5, 1, 3);
-
-        CommonClass.placements();
+        CommonSpawning.OVERWORLD_SPAWNS.forEach(spawnerData ->
+                BiomeModifications.addSpawn(BiomeSelectors.foundInOverworld(), MobCategory.MONSTER, spawnerData.type, spawnerData.getWeight().asInt(), spawnerData.minCount, spawnerData.maxCount)
+        );
+        CommonSpawning.NETHER_SPAWNS.forEach(spawnerData ->
+                BiomeModifications.addSpawn(BiomeSelectors.foundInTheNether(), MobCategory.MONSTER, spawnerData.type, spawnerData.getWeight().asInt(), spawnerData.minCount, spawnerData.maxCount)
+        );
+        CommonSpawning.SWAMP_SPAWNS.forEach(spawnerData ->
+                BiomeModifications.addSpawn(BiomeSelectors.includeByKey(Biomes.SWAMP, Biomes.MANGROVE_SWAMP), MobCategory.MONSTER, spawnerData.type, spawnerData.getWeight().asInt(), spawnerData.minCount, spawnerData.maxCount)
+        );
+        CommonSpawning.MOUNTAIN_SPAWNS.forEach(spawnerData ->
+                BiomeModifications.addSpawn(BiomeSelectors.includeByKey(Biomes.MEADOW, Biomes.FROZEN_PEAKS, Biomes.JAGGED_PEAKS, Biomes.STONY_PEAKS, Biomes.SNOWY_SLOPES, Biomes.CHERRY_GROVE), MobCategory.MONSTER, spawnerData.type, spawnerData.getWeight().asInt(), spawnerData.minCount, spawnerData.maxCount)
+        );
+        CommonSpawning.DEEP_DARK_SPAWNS.forEach(spawnerData ->
+                BiomeModifications.addSpawn(BiomeSelectors.includeByKey(Biomes.DEEP_DARK), MobCategory.MONSTER, spawnerData.type, spawnerData.getWeight().asInt(), spawnerData.minCount, spawnerData.maxCount)
+        );
+        CommonSpawning.END_SPAWNS.forEach(spawnerData ->
+                BiomeModifications.addSpawn(BiomeSelectors.foundInTheEnd(), MobCategory.MONSTER, spawnerData.type, spawnerData.getWeight().asInt(), spawnerData.minCount, spawnerData.maxCount)
+        );
+        CommonSpawning.placements();
     }
 }