+++ Forge/src/main/java/com/nyfaria/spookybats/event/CommonModEvents.java
@@ -1,6 +1,7 @@
 package com.nyfaria.spookybats.event;
 
 import com.nyfaria.spookybats.CommonClass;
+import com.nyfaria.spookybats.CommonSpawning;
 import com.nyfaria.spookybats.init.EntityInit;
 import net.minecraftforge.event.entity.EntityAttributeCreationEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -13,6 +14,6 @@ public class CommonModEvents {
     @SubscribeEvent
     public static void attribs(EntityAttributeCreationEvent e) {
         EntityInit.attributeSuppliers.forEach(p -> e.put(p.entityTypeSupplier().get(), p.factory().get().build()));
-        CommonClass.placements();
+        CommonSpawning.placements();
     }
 }